Time: 7 ms (21.77%), Space: 122.4 MB (7.26%) - LeetHub

+++ 2249-count-the-hidden-sequences/2249-count-the-hidden-sequences.cpp
@@ -12,8 +12,8 @@ public:
             prefix[i]=prefix[i-1]+differences[i];
         }
 
-        long long minEl=*min_element(prefix.begin(),prefix.end());
-        long long maxEl=*max_element(prefix.begin(),prefix.end());
+        int minEl=*min_element(prefix.begin(),prefix.end());
+        int maxEl=*max_element(prefix.begin(),prefix.end());
         
 
         int cnt=0;
